Clarify format of test.txt
Add "Required" and "Optional" elements lines underneath respective
elements' descriptions. Correct spelling and grammar in some places

The addition of the required and optional lines makes it easy to see,
at a glance, what is needed to build a given element, both for the JSON
text and the subsequent HTML formbuilding.